Make it easier to navigate to message macros in proto crate
* Move other infrequently-changed message traits to a separate file
* Rename the other proto.rs to message_stream.rs
